requirements.txt
langchain-groq
langchain-ollama
langchain-core
langchain-community
pypdf
python-dotenv
gradio
ollama